requirements.txt
# Core orchestration
langgraph>=0.2.0
langgraph-supervisor>=0.0.3
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0

# Legal NLP
spacy>=3.7.0
spacy-transformers>=1.3.0

# PDF / DOCX extraction
pdfplumber>=0.11.0
python-docx>=1.1.0

# Security / injection detection
pytector>=0.0.3

# UI
streamlit>=1.40.0

# Schema validation
pydantic>=2.0.0

# Structured logging
structlog>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
# unittest.mock is built into Python 3 — no separate install needed

# Environment
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.9.0
